# feat(crypto): reduce parameter copying and zeroing