```
$ benchstat -geomean old.txt new.txt
name     old time/op    new time/op    delta
Keys-12     401ns ± 2%     381ns ± 1%  -5.13%  (p=0.000 n=23+21)
```